выносит лоудеры роутов в папки со страницам
