--- packages/client/src/pages/app/app.module.scss
@@ -1,3 +1,0 @@
-.app {
-  color: green;
-}
