--- packages/client/src/pages/app/app.test.tsx
@@ -1,14 +1,0 @@
-import App from './app'
-import { render, screen } from '@testing-library/react'
-
-const appContent = 'Главная страница'
-
-// @ts-ignore
-global.fetch = jest.fn(() =>
-  Promise.resolve({ json: () => Promise.resolve('hey') })
-)
-
-test('Example test', async () => {
-  render(<App />)
-  expect(screen.getByText(appContent)).toBeDefined()
-})
