+++ packages/client/src/pages/game/game.tsx
@@ -1,3 +1,7 @@
+import { unAuthorizedRedirect } from '../../utils/helpers'
+
+export const gameLoader = unAuthorizedRedirect
+
 const Game = () => {
   return <div>Игра</div>
 }
