+++ packages/client/src/pages/index.ts
@@ -1,0 +1,23 @@
+import RootBoundary from './error-page/error-page'
+import Forum from './forum/forum'
+import Game, { gameLoader } from './game/game'
+import Home from './home/home'
+import Leaderboard from './leaderboard/leaderboard'
+import Login, { loginLoader } from './login/login'
+import Profile, { profileLoader } from './profile/profile'
+import SignUp, { signUpLoader } from './signup/signup'
+
+export {
+  RootBoundary,
+  Forum,
+  Game,
+  gameLoader,
+  Home,
+  Leaderboard,
+  Login,
+  loginLoader,
+  Profile,
+  profileLoader,
+  SignUp,
+  signUpLoader,
+}
