+++ packages/client/src/pages/login/login.tsx
@@ -1,3 +1,7 @@
+import { authorizedRedirect } from "../../utils/helpers"
+
+export const loginLoader = authorizedRedirect;
+
 const Login = () => {
   return <div>Страница авторизации</div>
 }
