+++ packages/client/src/pages/profile/profile.tsx
@@ -1,3 +1,7 @@
+import { unAuthorizedRedirect } from '../../utils/helpers'
+
+export const profileLoader = unAuthorizedRedirect
+
 const Profile = () => {
   return <div>Страница профиля</div>
 }
