+++ packages/client/src/pages/signup/signup.tsx
@@ -1,3 +1,7 @@
+import { authorizedRedirect } from '../../utils/helpers'
+
+export const signUpLoader = authorizedRedirect;
+
 const SignUp = () => {
   return <div>Страница регистрации</div>
 }
